requirements.txt
fastapi 
uvicorn 
transformers 
pillow 
requests 
torch